- Changed alot of doubles from float - Remove redundance from IModel - WIP

+++ RoungOut/core/src/Model/GameObjects/Pad.java
@@ -33,10 +33,10 @@ public class Pad {
 
 
     //Setters
-    public void setPadXPos(double padXPos) {
+    public void setPadXPos(float padXPos) {
         body.setX(padXPos);
     }
-    public void setPadYPos(double padYPos) {
+    public void setPadYPos(float padYPos) {
         body.setY(padYPos);
     }
     public void setSpeed(float speed) {
@@ -44,7 +44,7 @@ public class Pad {
     }
 
     //Constructor
-    public Pad(float length, float width, double padXPos, double padYPos,float padSpeed) {
+    public Pad(float length, float width, float padXPos, float padYPos,float padSpeed) {
         this.body =new RectangleBody(padXPos,padYPos,width,length);
         this.setSpeed(padSpeed);
     }
@@ -58,7 +58,7 @@ public class Pad {
     }
 
     //Overloaded Method: This simply sets the X and Y pos to the given values
-    public void padMove(double xPos, double yPos) {
+    public void padMove(float xPos, float yPos) {
         setPadXPos(xPos);
         setPadYPos(yPos);
     }
